+++ src/routes/Deriv/Help/index.js
@@ -9,11 +9,27 @@ import { derivInfo } from '../'
 
 const DerivHelp = props => (
   <Help toolInfo={derivInfo}>
-    <Notice>This tool is still in planning.</Notice>
+    <Notice>This tool is still under construction.</Notice>
     <h3 id='using'>
       Using {siteTitle}
       {derivInfo.title}
     </h3>
+    <h4 id='lexemes'>Lexemes</h4>
+    <p>
+      Enter a list of all lexemes that should be used as roots to be derived
+      from. Also add a brief definition for each of the lexemes.
+    </p>
+    <h4 id='derivations'>Derivational Affixes</h4>
+    <p>
+      Enter a list of all derivational affixes that should be used to derive new
+      words. Add a label for each of the affixes and a brief definition of the
+      change the affix creates.
+    </p>
+    <h4 id='derive'>Derive</h4>
+    <p>
+      When the list of lexemes and affixes is ready, click the Derive button to
+      generate a list of roots with possible derivational affixes added to them.
+    </p>
     <h3 id='acknowledgments'>Acknowledgments</h3>
     <p>
       Much thanks should be given to Jan Strasser and{' '}