+++ src/routes/Deriv/defaultData.js
@@ -1,0 +1,4 @@
+export const defData = {
+  words: 5,
+  type: 'any'
+}