+++ src/routes/Deriv/styles.js
@@ -1,3 +1,9 @@
-const styles = {}
+import * as v from '../../styles/variables'
+
+const styles = {
+  wordsInput: {
+    width: `${v.ms6}rem`
+  }
+}
 
 export default styles