+++ CHANGELOG.md
@@ -6,7 +6,15 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 
 ## [Unreleased](https://github.com/nai888/langua/compare/v0.2.5...HEAD)
 
+### Added 0.2.6
+
+- Add the Deriv form controls.
+- Add the Deriv default data.
+- Start adding instructions to Deriv Help.
+
 ### Changed 0.2.6
+
+- Update dependencies.
 
 ## [v0.2.5](https://github.com/nai888/langua/compare/v0.2.4...v0.2.5) - 2018-12-11